Modified the internal ReqRest serializers to throw NotSupportedExceptions. Added the ApiRequest.FetchAsync methods. Added the ApiRequest.ReceiveString methods. Added the ApiRequest.ReceiveByteArray methods. Updated the XML documentation.
