+++ src/ReqRest.Serializers/HttpContentSerializer.cs
@@ -1,7 +1,6 @@
 ﻿namespace ReqRest.Serializers
 {
     using System;
-    using System.Globalization;
     using System.Net.Http;
     using System.Text;
     using System.Threading.Tasks;
