+++ src/ReqRest.Tests/Internal/Serializers/StringSerializer/DeserializeTests.cs
@@ -1,12 +1,12 @@
-﻿namespace ReqRest.Tests.Internal.Serializers.NoContentSerializer
+﻿namespace ReqRest.Tests.Internal.Serializers.StringSerializer
 {
     using System;
-    using System.Net.Http;
     using System.Threading.Tasks;
-    using FluentAssertions;
     using ReqRest.Internal.Serializers;
-    using ReqRest.Serializers;
     using Xunit;
+    using FluentAssertions;
+    using ReqRest.Serializers;
+    using System.Net.Http;
 
     public class DeserializeTests
     {
@@ -14,17 +14,19 @@
         [Fact]
         public async Task Throws_NotSupportedException_When_Deserializing_Other_Type()
         {
-            var serializer = new NoContentSerializer();
+            var serializer = new StringSerializer();
             Func<Task> testCode = async () => await serializer.DeserializeAsync(new ByteArrayContent(Array.Empty<byte>()), typeof(int));
             (await testCode.Should().ThrowAsync<HttpContentSerializationException>()).WithInnerException<NotSupportedException>();
         }
 
         [Fact]
-        public async Task Deserializes_NoContent()
+        public async Task Deserializes_String()
         {
-            var serializer = new NoContentSerializer();
-            var deserialized = await serializer.DeserializeAsync(new ByteArrayContent(Array.Empty<byte>()), typeof(NoContent));
-            deserialized.Should().BeOfType<NoContent>();
+            var expected = "Hello World";
+            var serializer = new StringSerializer();
+            var deserialized = await serializer.DeserializeAsync(new StringContent(expected), typeof(string));
+            deserialized.Should().BeOfType<string>();
+            deserialized.Should().BeEquivalentTo(expected);
         }
 
     }
