+++ src/ReqRest/Internal/Serializers/NoContentSerializer.cs
@@ -4,7 +4,6 @@
     using System.Net.Http;
     using System.Text;
     using System.Threading.Tasks;
-    using ReqRest.Resources;
     using ReqRest.Serializers;
 
     /// <summary>
@@ -28,10 +27,10 @@
         // These two should never be called, because the base class handles the NoContent serialization.
         // If we ever get here, it was forced by the user through some tricks.
         protected override Task<object?> DeserializeCore(HttpContent httpContent, Type contentType) =>
-            throw new InvalidOperationException(ExceptionStrings.NoContentSerializer_CanOnlySerializeNoContentType());
+            throw new NotSupportedException();
 
         protected override HttpContent SerializeCore(object? content, Encoding encoding) =>
-            throw new InvalidOperationException(ExceptionStrings.NoContentSerializer_CanOnlySerializeNoContentType());
+            throw new NotSupportedException();
 
     }
